requirements/base.txt
docker-py==0.7.2
structlog==15.0.0
python-dateutil==2.4.0
